Rework resource mapping and polymorphism
Replaces `deepCast` with a generalized and well-tested `mapSchema` function that
traverses an object and a schema together and runs through a function. At the same
time, fix the `omitUndefined` function to properly recurse and build a number of
better tests for polymorphic resources.

Unfortunately, this logic required a bit of work around how to resolved `$ref`
types, which I chose to punt on until a new PR. So this version should not be
released yet.