Rework reference implementation.
References (`$ref`) as part of JSON Schema and OpenAPI and are necessary to implement
JSON Schema arrays and invaluable when implementing OpenAPI requests and responses.

However, OpenAPI 2.0 and 3.0.0 encode these references different (because they point
to different locations within the overall spec). As a result, we use a concrete `Reference`
class for all `$ref` instances and resolve it the time we wish to build an OpenAPI spec.

But: we need to resolve these references differently when performing resource validation
(e.g. without the OpenAPI spec). We do so by substituting the `Reference` with a placeholder
schema path and modifying the validator accordingly.