Simplify OpenAPI serving
We have been conflating building a spec (which needs to happen once)
with serving a spec (which needs to choose a version). The solution is
to enable the spec itself to act as the express route that serves itself.

In the process, generalize the version switching code to handle this case
in addition to the cases that build a spec into OpenAPI by version.